# perf(chunkenc): intern custom values for native histograms