Fixes: #13490 . Although we can do more if we implement not losing custom values in the iterator between iterator reuse.

```
$ benchstat nhcb-intern-baseline.txt nhcb-intern-pr.txt 
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Core(TM) Ultra 7 155U
                                                                          │ nhcb-intern-baseline.txt │        nhcb-intern-pr.txt         │
                                                                          │          sec/op          │   sec/op     vs base              │
NativeHistogramsCustomBuckets/sum-14                                                     1.260 ± 18%   1.157 ± 15%       ~ (p=0.132 n=6)
NativeHistogramsCustomBuckets/sum_rate_with_short_rate_interval-14                       1.232 ± 34%   1.422 ± 25%       ~ (p=0.589 n=6)
NativeHistogramsCustomBuckets/sum_rate_with_long_rate_interval-14                        2.204 ±  8%   2.194 ± 27%       ~ (p=0.818 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-14                1.185 ± 21%   1.194 ± 10%       ~ (p=1.000 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-14                 2.082 ± 14%   1.730 ± 27%       ~ (p=0.093 n=6)
geomean                                                                                  1.532         1.494        -2.45%

                                                                          │ nhcb-intern-baseline.txt │          nhcb-intern-pr.txt          │
                                                                          │           B/op           │     B/op       vs base               │
NativeHistogramsCustomBuckets/sum-14                                                    472.5Mi ± 1%   406.3Mi ±  2%  -14.03% (p=0.002 n=6)
NativeHistogramsCustomBuckets/sum_rate_with_short_rate_interval-14                      203.6Mi ± 4%   170.7Mi ±  0%  -16.19% (p=0.002 n=6)
NativeHistogramsCustomBuckets/sum_rate_with_long_rate_interval-14                       200.3Mi ± 0%   167.4Mi ±  0%  -16.46% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-14               148.7Mi ± 0%   147.1Mi ± 10%        ~ (p=0.394 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-14                153.0Mi ± 5%   157.9Mi ±  0%        ~ (p=1.000 n=6)
geomean                                                                                 213.0Mi        193.2Mi         -9.29%

                                                                          │ nhcb-intern-baseline.txt │         nhcb-intern-pr.txt         │
                                                                          │        allocs/op         │  allocs/op   vs base               │
NativeHistogramsCustomBuckets/sum-14                                                     6.589M ± 0%   5.160M ± 0%  -21.68% (p=0.002 n=6)
NativeHistogramsCustomBuckets/sum_rate_with_short_rate_interval-14                       3.719M ± 0%   2.998M ± 0%  -19.37% (p=0.002 n=6)
NativeHistogramsCustomBuckets/sum_rate_with_long_rate_interval-14                        3.611M ± 0%   2.891M ± 0%  -19.95% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-14                1.726M ± 0%   1.690M ± 3%        ~ (p=0.065 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-14                 1.633M ± 4%   1.611M ± 0%   -1.34% (p=0.002 n=6)
geomean                                                                                  3.015M        2.613M       -13.36%
```