@krajorama 2025-05-07
It is an open question whether to keep counting the size of the custom values for the engine [here](https://github.com/prometheus/prometheus/blob/2bf6f4c9dcbb1ad2e8fef70c6a48d8fc44a7f57c/model/histogram/float_histogram.go#L521).